requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
docker-pycreds==0.4.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.12.0
gitdb==4.0.11
GitPython==3.1.40
hydra-core==1.3.2
idna==3.6
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.2
lightning==2.1.2
lightning-utilities==0.10.0
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
omegaconf==2.3.0
packaging==23.2
pandas==2.1.3
protobuf==4.25.1
psutil==5.9.6
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==2.1.2
pytz==2023.3.post1
PyYAML==6.0.1
rdflib==7.0.0
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
sentry-sdk==1.38.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
torch==2.1.1
torch-cluster==1.6.3
torch-scatter==2.1.2
torch-sparse @ git+https://github.com/rusty1s/pytorch_sparse.git@05b62b1c02447f28f87dfe2d44ae8d8d2e0204ba
torch-spline-conv==1.2.2
torch_geometric==2.4.0
torchmetrics==1.2.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wandb==0.16.1
yarl==1.9.3
